Unify CirclesScreen.cs across MonoGame/raylib/SilkNet samples
Converges the three drifted per-backend CirclesScreen.cs copies into one shared file (MonoGameGumInCode is canonical), following the same pattern already used for TextScreen.cs/RenderTargetScreen.cs. Linked into GumTest.csproj and SilkNetGumSample.csproj via <Compile Include Link>.

Colors are built via new Color(r,g,b,a) instead of named statics since SKColor has no static named-color members. The Antialiasing section is excluded on raylib (no per-shape AA) and the Blend section is excluded on Skia (no Blend on SkiaShapeRuntime); both were already backend-specific gaps. The Rotation (outline) row substitutes a "not supported" label on raylib (LineCircle has no gradient-on-stroke path) instead of being omitted, keeping section order aligned.

Closes #4024
