+++ Samples/raylib/GumTest.csproj
@@ -18,13 +18,14 @@
     <ProjectReference Include="..\..\Integrations\KernSmith\KernSmith.RaylibGum\KernSmith.RaylibGum.csproj" />
   </ItemGroup>
 
-  <!-- #3640/#3988/#3998: TextScreen.cs, RenderTargetScreen.cs, and RenderTargetShaderScreen.cs are
-       shared with Samples/MonoGameGumInCode (the canonical copies), gated #if RAYLIB, to stop the
-       mirrors drifting when only one got updated. -->
+  <!-- #3640/#3988/#3998/#4024: TextScreen.cs, RenderTargetScreen.cs, RenderTargetShaderScreen.cs,
+       and CirclesScreen.cs are shared with Samples/MonoGameGumInCode (the canonical copies), gated
+       #if RAYLIB, to stop the mirrors drifting when only one got updated. -->
   <ItemGroup>
     <Compile Include="..\MonoGameGumInCode\MonoGameGumInCode\Screens\TextScreen.cs" Link="Screens\TextScreen.cs" />
     <Compile Include="..\MonoGameGumInCode\MonoGameGumInCode\Screens\RenderTargetScreen.cs" Link="Screens\RenderTargetScreen.cs" />
     <Compile Include="..\MonoGameGumInCode\MonoGameGumInCode\Screens\RenderTargetShaderScreen.cs" Link="Screens\RenderTargetShaderScreen.cs" />
+    <Compile Include="..\MonoGameGumInCode\MonoGameGumInCode\Screens\CirclesScreen.cs" Link="Screens\CirclesScreen.cs" />
   </ItemGroup>
 
   <ItemGroup>
